Update mobile styling in home page
